Move spec to service
