+++ airbyte-ci/connectors/metadata_service/lib/pyproject.toml
@@ -10,6 +10,7 @@ packages = [{include = "metadata_service"}]
 python = "^3.9"
 pydantic = "^1.10.6"
 click = "^8.1.3"
+google = "^3.0.0"
 
 
 [tool.poetry.group.test.dependencies]
