+++ airbyte-ci/connectors/metadata_service/orchestrator/orchestrator/assets/metadata_assets.py
@@ -1,9 +1,8 @@
 import pandas as pd
 from dagster import MetadataValue, Output, asset
-from deepdiff import DeepDiff
-import mergedeep
 
 from metadata_service.models.generated.ConnectorMetadataDefinitionV1 import ConnectorMetadataDefinitionV1
+from orchestrator.utils.object_helpers import are_values_equal, merge_values
 
 GROUP_NAME = "metadata"
 
@@ -12,21 +11,6 @@ GROUP_NAME = "metadata"
 
 OSS_SUFFIX = "_oss"
 CLOUD_SUFFIX = "_cloud"
-
-def are_values_equal(value_1, value_2):
-    if isinstance(value_1, dict) and isinstance(value_2, dict):
-        diff = DeepDiff(value_1, value_2, ignore_order=True)
-        return len(diff) == 0
-    else:
-        return value_1 == value_2
-
-def merge_values(old_value, new_value):
-    if isinstance(old_value, dict) and isinstance(new_value, dict):
-        merged = old_value.copy()
-        mergedeep.merge(merged, new_value)
-        return merged
-    else:
-        return new_value
 
 
 def get_primary_catalog_suffix(merged_df):
